leads permissions

+++ leads/views.py
@@ -61,7 +61,7 @@ class LeadListView(APIView, LimitOffsetPagination):
                 "assigned_to",
             )
         ).order_by("-id")
-        if self.request.profile.role != "ADMIN" and not self.request.user.is_superuser:
+        if self.request.profile.role not in ["ADMIN", "SALES MANAGER","SALES REP"] and  not self.request.user.is_superuser:
             queryset = queryset.filter(
                 Q(assigned_to__in=[self.request.profile])
                 | Q(created_by=self.request.profile.user)
@@ -273,7 +273,7 @@ class LeadDetailView(APIView):
         ]
         if self.request.profile.user == self.lead_obj.created_by:
             user_assgn_list.append(self.request.profile.user)
-        if self.request.profile.role != "ADMIN" and not self.request.user.is_superuser:
+        if self.request.profile.role not in ["ADMIN", "SALES MANAGER","SALES REP"] and not self.request.user.is_superuser:
             if self.request.profile.id not in user_assgn_list:
                 return Response(
                     {
@@ -371,7 +371,7 @@ class LeadDetailView(APIView):
                 {"error": True, "errors": "User company doesnot match with header...."},
                 status=status.HTTP_403_FORBIDDEN,
             )
-        if self.request.profile.role != "ADMIN" and not self.request.user.is_superuser:
+        if self.request.profile.role not in ["ADMIN", "SALES MANAGER"] and not self.request.user.is_superuser:
             if not (
                 (self.request.profile.user == self.lead_obj.created_by)
                 or (self.request.profile in self.lead_obj.assigned_to.all())
@@ -423,8 +423,23 @@ class LeadDetailView(APIView):
                     "error": True,
                     "errors": "User company does not match with header....",
                 },
-                status=status.HTTP_403_FORBIDDEN,
-            )
+                 status=status.HTTP_403_FORBIDDEN,
+                )
+                           
+        if self.request.profile.role not in ["ADMIN", "SALES MANAGER","SALES REP"] and not self.request.user.is_superuser:
+            if not (
+                (self.request.profile.user == self.lead_obj.created_by)
+                or (self.request.profile in self.lead_obj.assigned_to.all())
+            ):
+                return Response(
+                    {
+                        "error": True,
+                        "errors": "You do not have Permission to perform this action",
+                    },
+                    status=status.HTTP_403_FORBIDDEN,
+                )
+                
+            
         serializer = LeadCreateSerializer(
             data=params,
             instance=self.lead_obj,
@@ -546,7 +561,7 @@ class LeadDetailView(APIView):
     def delete(self, request, pk, **kwargs):
         self.object = self.get_object(pk)
         if (
-            request.profile.role == "ADMIN"
+            self.request.profile.role not in ["ADMIN", "SALES MANAGER"] 
             or request.user.is_superuser
             or request.profile.user
              == self.object.created_by
@@ -601,7 +616,7 @@ class LeadCommentView(APIView):
         params = request.data
         obj = self.get_object(pk)
         if (
-            request.profile.role == "ADMIN"
+            self.request.profile.role not in ["ADMIN", "SALES MANAGER","SALES REP"] 
             or request.user.is_superuser
             or request.profile == obj.commented_by
         ):
@@ -628,7 +643,7 @@ class LeadCommentView(APIView):
     def delete(self, request, pk, format=None):
         self.object = self.get_object(pk)
         if (
-            request.profile.role == "ADMIN"
+            request.profile.role == ["ADMIN", "SALES MANAGER"]
             or request.user.is_superuser
             or request.profile == self.object.commented_by
         ):
@@ -656,7 +671,7 @@ class LeadAttachmentView(APIView):
     def delete(self, request, pk, format=None):
         self.object = self.model.objects.get(pk=pk)
         if (
-            request.profile.role == "ADMIN"
+            self.request.profile.role not in ["ADMIN", "SALES MANAGER"] 
             or request.user.is_superuser
             or request.profile.user == self.object.created_by
         ):
@@ -809,23 +824,47 @@ class CompanyDetail(APIView):
     @extend_schema(tags=["Company"],description="Company Update",parameters=swagger_params1.organization_params,request=CompanySerializer)
     def put(self, request, pk, format=None):
         company = self.get_object(pk)
-        serializer = CompanySerializer(company, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(
-                {"error": False, "data": serializer.data,'message': 'Updated Successfully'},
-                status=status.HTTP_200_OK,
+        if (
+            self.request.profile.role not in ["ADMIN", "SALES MANAGER","SALES REP"] 
+            or request.user.is_superuser
+        ):
+            serializer = CompanySerializer(company, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(
+                        {"error": False, "data": serializer.data,'message': 'Updated Successfully'},
+                        status=status.HTTP_200_OK,
             )
-        return Response(
+            return Response(
                 {"error": True,'message': serializer.errors},
                 status=status.HTTP_400_BAD_REQUEST,
-            )
+                        )
+        return Response(
+            {
+                "error": True,
+                "errors": "You don't have permission to perform this action",
+            },
+            status=status.HTTP_403_FORBIDDEN,
+        )
+
     @extend_schema(tags=["Company"],parameters=swagger_params1.organization_params)
     def delete(self, request, pk, format=None):
         company = self.get_object(pk)
-        company.delete()
-        return Response(
-                {"error": False, 'message': 'Deleted successfully'},
+        if (
+            request.profile.role == ["ADMIN", "SALES MANAGER"]
+            or request.user.is_superuser
+            or request.profile == self.object.commented_by
+        ):
+            company.delete()
+            return Response(
+                {"error": False, "message": "Deleted Successfully"},
                 status=status.HTTP_200_OK,
             )
- 
+
+        return Response(
+            {
+                "error": True,
+                "errors": "You do not have permission to perform this action",
+            },
+            status=status.HTTP_403_FORBIDDEN,
+        )
